# Conflicts:
#	app/build.gradle

+++ app/src/main/java/eu/faircode/email/WorkerCleanup.java
@@ -39,6 +39,7 @@ import java.util.concurrent.TimeUnit;
 
 public class WorkerCleanup extends Worker {
     private static final int CLEANUP_INTERVAL = 4; // hours
+    private static final long KEEP_FILES_DURATION = 3600 * 1000L; // milliseconds
     private static final long CACHE_IMAGE_DURATION = 3 * 24 * 3600 * 1000L; // milliseconds
     private static final long KEEP_CONTACTS_DURATION = 180 * 24 * 3600 * 1000L; // milliseconds
     private static final long KEEP_LOG_DURATION = 24 * 3600 * 1000L; // milliseconds
@@ -110,15 +111,16 @@ public class WorkerCleanup extends Worker {
 
             // Cleanup message files
             Log.i("Cleanup message files");
-            for (File file : files) {
-                long id = Long.parseLong(file.getName().split("\\.")[0]);
-                EntityMessage message = db.message().getMessage(id);
-                if (message == null || !message.content) {
-                    Log.i("Deleting " + file);
-                    if (!file.delete())
-                        Log.w("Error deleting " + file);
+            for (File file : files)
+                if (manual || file.lastModified() + KEEP_FILES_DURATION < now) {
+                    long id = Long.parseLong(file.getName().split("\\.")[0]);
+                    EntityMessage message = db.message().getMessage(id);
+                    if (message == null || !message.content) {
+                        Log.i("Deleting " + file);
+                        if (!file.delete())
+                            Log.w("Error deleting " + file);
+                    }
                 }
-            }
 
             // Cleanup attachment files
             Log.i("Cleanup attachment files");
@@ -164,21 +166,30 @@ public class WorkerCleanup extends Worker {
     }
 
     static void queue(Context context) {
-        Log.i("Queuing " + getName() + " every " + CLEANUP_INTERVAL + " hours");
+        try {
+            Log.i("Queuing " + getName() + " every " + CLEANUP_INTERVAL + " hours");
 
-        PeriodicWorkRequest workRequest =
-                new PeriodicWorkRequest.Builder(WorkerCleanup.class, CLEANUP_INTERVAL, TimeUnit.HOURS)
-                        .build();
-        WorkManager.getInstance(context)
-                .enqueueUniquePeriodicWork(getName(), ExistingPeriodicWorkPolicy.REPLACE, workRequest);
+            PeriodicWorkRequest workRequest =
+                    new PeriodicWorkRequest.Builder(WorkerCleanup.class, CLEANUP_INTERVAL, TimeUnit.HOURS)
+                            .build();
+            WorkManager.getInstance(context)
+                    .enqueueUniquePeriodicWork(getName(), ExistingPeriodicWorkPolicy.REPLACE, workRequest);
 
-        Log.i("Queued " + getName());
+            Log.i("Queued " + getName());
+        } catch (IllegalStateException ex) {
+            // https://issuetracker.google.com/issues/138465476
+            Log.w(ex);
+        }
     }
 
     static void cancel(Context context) {
-        Log.i("Cancelling " + getName());
-        WorkManager.getInstance(context).cancelUniqueWork(getName());
-        Log.i("Cancelled " + getName());
+        try {
+            Log.i("Cancelling " + getName());
+            WorkManager.getInstance(context).cancelUniqueWork(getName());
+            Log.i("Cancelled " + getName());
+        } catch (IllegalStateException ex) {
+            Log.w(ex);
+        }
     }
 
     private static String getName() {
